+++ src/tko/play/search.py
@@ -24,76 +24,64 @@ class Search:
         if self.search_mode:
             self.backup_expanded = [v for v in self.tree.expanded]
             self.backup_index_selected = self.tree.selected_item
-            self.backup_admin_mode = Flags.admin.is_true()
+            self.backup_admin_mode = Flags.admin
             self.tree.update_tree(admin_mode=True)
             self.tree.process_expand()
             self.tree.process_expand()
-            self.fman.add_input(Floating(">v").warning().put_text("Digite o texto\nNavegue até o elemento desejado\ne aperte Enter"))
+            #self.fman.add_input(Floating(">v").warning().put_text("Digite o texto\nNavegue até o elemento desejado\ne aperte Enter"))
     
     def finish_search(self):
+        if self.tree.selected_item == "":
+            self.cancel_search()
+            return
+
         self.search_mode = False
-        unit = self.tree.get_selected()
-        self.tree.selected_item = ""
         self.tree.search_text = ""
-        if self.backup_admin_mode == False:
-            self.tree.update_tree(admin_mode=False)
+        selected_key = self.tree.selected_item
+        self.tree.update_tree(self.backup_admin_mode) # usa o mode de antes e vê se acha
         self.tree.reload_sentences()
     
         found = False
-        for i, item in enumerate(self.tree.items):
-            if item == unit:
-                self.tree.selected_item = item.get_key()
+        for unit in self.tree.items:
+            if unit.key == selected_key:
                 found = True
                 break
 
         if not found:
-            self.fman.add_input(Floating(">v").warning().put_text("Elemento não acessível no modo normal.\nEntrando no modo Admin\npara habilitar acesso"))
+            # self.fman.add_input(Floating(">v").warning().put_text("Elemento não acessível no modo normal.\nEntrando no modo Admin\npara habilitar acesso"))
             Flags.admin.toggle()
             self.tree.update_tree(True)
             self.tree.reload_sentences()
         
-        self.tree.process_collapse()
-        self.tree.process_collapse()
-        self.tree.selected_item = unit.get_key()
+        self.tree.expanded = []
+        unit = self.tree.all_items[selected_key]
+        with open("log.txt", "a") as f:
+                f.write(f"Selected: {selected_key} {type(unit)} \n")
+
         if isinstance(unit, Task):
-            for cluster_key in self.game.available_clusters:
-                cluster = self.game.clusters[cluster_key]
-                for quest in cluster.quests:
-                    for task in quest.get_tasks():
-                        if task == unit:
-                            self.tree.expanded = [cluster.key, quest.key]
+            with open("log.txt", "a") as f:
+                f.write(f"Selected: {selected_key} {unit.cluster_key} {unit.quest_key} \n")
+            self.tree.expanded = [unit.cluster_key, unit.quest_key]
         elif isinstance(unit, Quest):
-            for cluster_key in self.game.available_clusters:
-                cluster = self.game.clusters[cluster_key]
-                for quest in cluster.quests:
-                    if quest == unit:
-                        self.tree.expanded = [cluster.key]
+            self.tree.expanded = [unit.key, unit.cluster_key]
         self.tree.reload_sentences()
-        for i, item in enumerate(self.tree.items):
-            if item== unit:
-                self.tree.selected_item = item.get_key()
-                break
 
     # update index to match the first item that matches the search
     def update_index(self):
-        filtered, first = self.tree.filter_by_search()
-        self.tree.selected_item = first if first is not None else self.tree.selected_item
-        # if first is not None:
-        #     for i, item in enumerate(self.tree.items):
-        #         if item.get_key() == first:
-        #             self.tree.selected_item = item.get_key()
-        #             break
+        _, first = self.tree.filter_by_search()
+        self.tree.selected_item = first if first is not None else ""
 
-
-    def process_search(self, key):
-        if key == 27:
+    def cancel_search(self):
             self.search_mode = False
             self.tree.search_text = ""
             self.tree.expanded = [v for v in self.backup_expanded]
             self.tree.selected_item = self.backup_index_selected
+
+    def process_search(self, key):
+        if key == InputManager.esc:
+            self.cancel_search()
         elif key == ord("\n"):
             self.finish_search()
-    
         elif key == curses.KEY_UP:
             self.tree.move_up()
         elif key == curses.KEY_DOWN: